backend/requirements.txt
asgiref==3.10.0
django-filter==24.3
Django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-spectacular==0.27.2
Pillow==11.0.0
psycopg2-binary==2.9.9
sqlparse==0.5.3
tzdata==2025.2
